Deduplicate models list used to dump/load H2 dbs
The failing load-from-h2 test that checks that all models are accounted
for alerted me that I need to handle the new `NativeQuerySnippet` model.
However, there was no such test for dump-to-h2. Instead of adding such a
test, I opted to simply use load-from-h2's list in dump-to-h2 to
eliminate the problem, rather than testing for it.