Allow / in expression names
Fixes metabase/metabase#12305. Using (name) will drop the namespace of a
Clojure keyword. During JSON parsing, a key like "test/foo" will end up
in the Clojure namespace of `test` with a keyword name of `foo`. In
these cases, we need to use the `qualified-name` function to properly
output the expression name.

[ci drivers] so that all drivers are tested