Fixing Google Analytics failures on new instances
This stops a failure during fingerprinting of a table from causing the
entire sync to fail.

Fingerprinting is now skipped for Google Analytics datasources, as they
do not support this (all queries against GA require a `metrics`
parameter)

Add an assert to ensure that `:metrics` is present in a GA query. The
Google library will fail with a NPE when it's not.

Resolves #12411

[ci googleanalytics]