First time sync BigQuery data w/ Service Account
When using a service account, the project-id isn't provided by the user,
but instead comes from the credentials for the service account. Handle
this case by looking for a project-id if one is provided explicitly,
otherwise defer to the credential object.

Additional change - during development, if (dev/init!) isn't called, the
event listeners are never setup, meaning that the initial syncs don't
occur.

Resolves #12648

[ci bigquery]