Allow connection URIs for Mongo (#12667)
Allows a connection URI string to be used for connecting to Mongo. When
using a connection URI you cannot provide the SSL certificate in a field on
settings page. This is a limitation of the underlying driver.

The conn-uri becomes a password field after it's been saved to avoid
displaying a password in cleartext.

Resolves #5084

[ci mongo]