Adds a is-empty/not-empty for strings
What the UI calls "is empty" / "not empty" right now is a null check,
not a null and "" string check.

This adds is-empty / not-empty operators in the query processor, and
reworks the UI to use those. Any existing query that previously used the
is-null / not-null operators will show (as "Is null" / "Not null") but
any newly made queries will not be able to select "Is null" / "Not null"
for text type fields.

This will allow backwards compatibility with existing questions, while
promoting use of the new "is-empty"/"not-empty" operators.

Resolves #5299