Dataset test tweaks (#13040)
* Wait a moment for query execution record to come in

It was observed when testing against AWS RDS/Aurora PostgreSQL, that
tests fail because it was comparing against the _second_ most recent
query.  I'm not sure why this happened, but the delay fixed it.

* Initialize db in test fixture

Running `lein test metabase.api.dataset-test`, some tests fail because
the db isn't yet initialized.

* Fix ns decl

* Select most recent query execution by start date, rather than ID

This is to get rid of the `Thread/sleep`.