Handle date formatting specifically for Excel
Excel will not parse an ISO8601 date, so add a new version of
FormatValue specifically made for Excel.

The Docjure library used has formatting helpers for a few Java classes,
but specifically java.util.Date and not the newer Java 8 Time classes.
In this case, it formats any java.util.Date as m/d/yy.

Resolves #10803