When a user specifies SSL explicitly, never try without it
Previously, when adding a database, we would try to connect to the
database once with SSL, and once without. If connecting without SSL
works, but connecting with did not, we save the database as not needing
SSL. This ignored the SSL setting the user specified.

This makes the user's setting take precedence in only allowing SSL
connections.

Resolves #13295