Half of fix for OOM in sync (#13288) (2.1) (#13325)
* Half of fix for OOM in sync (#13288) (2.1)

* truncate type/Text fields to 1234 (a distinct number we can use as a
heuristic for longer fields if desired)

* test truncation in metadata-queries-test

* test that fingerprints reflect these truncations in fingerprint-test

* Check database_type of text for substring of summaries #13288

in pg, many types go to :type/Text which cannot have substring on
them, such as xml, json, jsonb, etc.

* Skip fingerprinting "structured" fields, take substring of text fields

* "Sniff" JSON results when truncated in fingerprinting

- had to move the truncation-size var due to circularity, but it most
logically belongs in the fingerprinting namespace as that's the only
mechanism that really should be aware of it anyways

- get rid of :refer :all in a test ns

- fingerprinting doesn't happen on base_type :type/Structured so add
that to the query for fields to fingerprint

* Move `table-rows-sample` truncation option into options map

* Switch json fingerprinting entirely to heuristic rather than parsing

- faster in all cases and accuracy is not a particular goal
- doesn't rely on the truncation size for switching to heuristic

* Derive SerializedJson and XML from :type/Structured

* Fixes metadata queries test when no truncation-size provided

* JSON heuristic for array of booleans

- want to validate as JSON [true, false] but not [random tag]

* Removes two unused deps from fingerprinters

we no longer actually parse json so don't need cheshire and don't use
clojure.string/starts-with in favor of a little descent parser

* CamelCases schema name and move truncation-size to appropriate ns

truncation size used to be used in the fingerprinting analysis when
checking text for JSON (using a heuristic if it was truncated) and
passed into the query to get table summaries. Now its only passed into
the summaries query and the JSON recognizer in the text fingerprinter
always uses heuristics

* Fix test now that its not relying on truncation limit for heuristics

previously was making a string that was the `truncation-limit` length
to trigger the heuristics of json parsing. now we always use heuristic
so no need to construct such a precise length string.