Only get substrings in fingerprinting when supported [ci drivers] (#13432)
* Only get substrings in fingerprinting when supported [ci drivers]

failing tests in bigquery and mongo as they don't support
expressions.

* Corrects ns form and gets driver from table def [ci drivers]

getting driver from dynamic var was not reliable at this place

* Move expression check into `metadata-queries/table-row-sample`

debated whether caller should know that whether to pass in truncation
options or they might "silently" be ignored. Probably best this way
and added that information to docstring.