Handle mysql YEAR column types (Fixes #13203) [ci drivers]
- year columns are very small in the db. The tricky part is how the
driver hands them back: either as shorts or as java.sql.Dates. Our
standard trick of reading them right into the java.time.LocalDate
doesn't work and we instead get the java.sql.Date (which the driver
converts from the short internal representation) or the short if
`yearIsDateType` is configured to false.

It seems to be an open question if Metabase should always consume this
as a Date. That seems most likely but I have no idea how often this
comes up in practice.