Handle mysql YEAR column types (Fixes #13203) [ci drivers] (#13448)
* Handle mysql YEAR column types (Fixes #13203) [ci drivers]

- year columns are very small in the db. The tricky part is how the
driver hands them back: either as shorts or as java.sql.Dates. Our
standard trick of reading them right into the java.time.LocalDate
doesn't work and we instead get the java.sql.Date (which the driver
converts from the short internal representation) or the short if
`yearIsDateType` is configured to false.

It seems to be an open question if Metabase should always consume this
as a Date. That seems most likely but I have no idea how often this
comes up in practice.

* Add typehints [ci drivers]

* Consume metadata immediately when checking for YEAR mysql column

don't want to access the metadata outside of a `jdbc/with-db-metadata`

* Clean up logic for YEAR column types [ci drivers]

the `:else` is actually bad syntax. And handle explicitly the nil case
and the two types it can be from getObject

* Mark YEAR as :type/Date, make tests more reliable

column ordering was messing up testing, so ignore the PK and just
select the YEAR column

* Use `when-let` when checking YEAR column [ci drivers]

* Add tests for `table-rows-sample`

really just to appease the Book of Cam® for test coverage