Smarter state fields (#13512)
* Mark state fields by fingerprint fixes #2735 [ci drivers]

* fix tests for pg, ns sort, and docstring lint

* Expectations -> clojure.test

* minor alignment in test

* Allow fingerprint classifiers to override prev classifier special_type

if another classifier threw on a special type due to name or the
number of items in the collection, this should override it as it is
introspecting values.

States previously were marked ":type/State" based on the column
name. Now use a fingerprinter looking for state names. Owing to the
fact that there might be more things in the state column (US
territory, canadian states, Washington DC, etc) lowered the threshold
for recognizing a state column from 95% of values down to 70% of
values.

* Cleanup [ci drivers]

- remark about original field in metadata
- docstring on `can-edit-special-field?`
- cleanup and check for string in `state?`
- add tests that nils and other types don't break `state?`

* Include test file