Testing that pagination occurs
Due to lack of metrics on calls made, a dynamic callback was added so
that call counts can be tracked to indicate that pagination had
occurred.

Also, introduces a maximum row count to each BigQuery call. Leaving at
an arbitrarily high number - 20,000. BigQuery will automatically stop at
a max of 10 MB per page, regardless of this number, but the setting is
used in tests to ensure what pagination is doing.

[ci bigquery]