Add pagination within the BigQuery driver (#13530)
* Add pagination within the BigQuery driver

The BigQuery API has a specific method of pagination using a "page
token" in a query result. Use this to retrieve the entire result set in
a lazy fashion.

Resolves #6960

[ci bigquery]

* Fixed tests to work with new interface

Added ability to use a service account json for BigQuery testing in
development

Still need to add a test specifically for pagination

[ci bigquery]

* Testing that pagination occurs

Due to lack of metrics on calls made, a dynamic callback was added so
that call counts can be tracked to indicate that pagination had
occurred.

Also, introduces a maximum row count to each BigQuery call. Leaving at
an arbitrarily high number - 20,000. BigQuery will automatically stop at
a max of 10 MB per page, regardless of this number, but the setting is
used in tests to ensure what pagination is doing.

[ci bigquery]

* Clarify max-results as being per page

indentation and other small review feedback

[ci bigquery]