Add refingerprint column to Database
its nullable now so that people can opt in and we can migrate to opt
out in the future with the following strategy:  if null, set to True,
set default to True. This allows us to respect people who have turned
it off and enabling in a future release when we are sure the
performance ramifications are not too severe.