Allow SSL client certificates to work in pgsql
When a user specifies a `sslmode` or `sslfactory`, we should honor that instead of overwriting their selections.

The NonValidatingFactory isn't something we should default to - it just checks that an SSL certificate is present, it does not attempt
any sort of validation of said certificate, and it also explicitly blocks support for client certificates.

Resolves metabase/metabase#13796

[ci postgres]