Remove object count from log output (#13823)
`count` caused the lazy `objs` to be realized, causing the entire table
to be scanned. For tables with very high numbers of rows this stalls the
program and effectively eliminates the usefulness of chunking `objs`.

If the count is _really_ that big a deal, it can be determined, in
multiples of `chunk-size`, but counting the number of dots after the log
line, as printed by `insert-chunk!`.