Don't set Site URL on requests from ELB HealthChecker (#12996)
* Fix for AWS ELB issue

When Metabase is deployed to AWS behind ELB the very first call to it is made by ELB health check. Out of `origin`, `x-forwarded-host`,  `host` in this case only `host` is populated and it is populated with internal private IP. This causes `site-url` to be populated with URL with private IP in it. The fix is checks for caller user agent and if it contains "HealthChecker" - does not set `site-url` at this time

* Added missing user-agent

* Added check for nil user-agent