Allow filters to work in pivot tables
The `query->expected-cols` function only worked with queries of the type `:query` but `"query"` is perfectly valid also.

Modify the function to take that type, and remove the extraneous calls to `preprocess-query` - the result of a call to
`preprocess-query` cannot be passed to `preprocess-query` a second time without causing a schema violation error.