13-sql-parameters.md: make backtick requirement more apparent
This is intended to make it less likely to miss the requirement that Filter Fields won't
work with BigQuery if backticks are not included in the dataset.tablename. It's possible
when reading quickly to read the first two sentences of the bullet and think your BQ
query is fine because it has the dataset name, but the third sentence additionally calls
out the backtick requirement. By separating that as a separate bullet, hopefully it will
be harder to miss.