Fix test on h2 db
When testing with h2 driver as the destination db, special case the
jdbc spec.

H2 being in memory makes that the usual jdbc connection that would be
created would be ephemeral, so when doing the dump, and comparing
again against the first, the db would be gone. Adding a CLOSE_DELAY=10
makes it stick for a little while so we can do the comparison.

not calling `tx/create-db!` in that case because in h2, db is always
created already.