Default databases with auto_run_queries to true (#14355)
* Default databases with auto_run_queries to true

previously this was set in the
frontend/src/metabase/entities/databases/forms.js file with initial:
true but this did not affect form state, just the initial value of the
form. So what happened was we saved nil and regardless of the value,
the initial value of the form was checked. Meaning if you saved it as
false, subsequent page loads would show it as true due to the initial
form state. The db has a default of true for this column which is why
this seemingly worked in the past

* Add repro for #14334

Co-authored-by: Nemanja <31325167+nemanjaglumac@users.noreply.github.com>