Fix schedule picker for pulses (#14413)
* Fix schedule picker for pulses

don't know why there's a `_.pick` here. This would have worked "out of
the box" if this didn't have opinions about which parts of the
schedule to throw away

* Also fix schedule picker for new dashboard subscriptions sent via Slack

* Revert "Fix schedule picker for pulses"

This reverts commit d10eb2727f95c67be56ef50d879c071f7c57482f.

* Make minute picker on scheduler optional

only needed (as yet) on the db schedule page. pulses don't use cron
strings but have each part of the schedule map as a db field so would
need a migration to use it.

* Revert "Also fix schedule picker for new dashboard subscriptions sent via Slack"

This reverts commit 05ed9da535e897e859c8f258b23f0892e15b2f57.

Co-authored-by: Tim Macdonald <tsmacdonald@gmail.com>