When using FieldValues, correct use cached values
the human-readable-values is a sequence of values or empty. If
non-empty, it can be zipmapped with the values to get value->hrv
map. But the lack of HRVs does NOT imply that we don't have cached
values to be used. If not, we just zipmap the original values with the
str'd version of themselves