Incorporate feedback from cam about macros
- it is super important that the db/table-syncs are off thread and the
request is not blocked. Getting that macro wrong would be devastating
and it only exists so it can be written `(execute (db-sync-fn
database))` instead of *gasp* `(execute #(db-sync-fn database))`. Used
in three places, and getting it wrong is terrible. Delaying into the
thunk makes clear that this will work just fine.