Remove the relation type migration.
it makes the diff far too hard at the moment. need to be able to just
move everything from special_type -> semantic_type and then correct
the few that look at it to care about the effective type.

the PK/FK stuff gets really invasive and needs to be in its own much
smaller changeset