Ensure connection string goes through jdbc without changes
we thought we were doing this but `(jdbc/get-connection "your
string")` still parses the connection string into its db-spec style
map. This seems innocuous but it doesn't url decode the uri so
passwords in particular don't work correctly.

The solution is to use `(jdbc/get-connection {:connection-uri "your
string"}})` This allows the dump/read to h2 stuff to use this as is
and just required a slight modification to the way we create the db
pool for the mb db.

This change is a bit larger but its mostly docstrings and a few tests
to make sure we're doing the right thing.