test fit CollectionList [ci skip]
trying out our existing CollectionList in the new SavedQuestionPicker to display collections.
good news is it looks like some of it will work, but we'll need to alter a few things that are
too specific to the initial case. but this seems like a reasonable avenue to continue prototyping with.