Use ReactDOMServer to render html in a contenteditable div (#14869)
As part of the react v16 we can no longer rely on a synchronous
ReactDOM.render call. Using the third arg callback of ReactDOM.render
does not work for reasons unknown to me, but it would likely be
race-condition-y anyways due to this being an input field. I'm using
ReactDOMServer in order to avoid rewriting the react TokenizedExpression
component as html and to avoid associated xss vulnerabilities, etc.