feat(serialization) do not dump :archived entities by default
Add an option `dump` command `--state (active|all)` defaulting to
active that filters the dumped entities to only the active ones.

This reduces the scope of another issue there is atm, where in case
there are multiple entitites with same names, only one survives the
dump. In many occasions, users archive a Card to create a "new one"
which is named equally. In that case, The archived one might be the
one that survives, leaving the user with 0 Cards after load.

This PR adds the default, and the cli flag, so we use a sane default,
reducing the chances of confusion.