Prevent delay when showing the expression suggestion for the first time
The delay was a minor regression caused by deferring the construction of
the expensive parser (see PR #14133) to prevent atrocious embedding
performance.  However, that also meant that the parser and everything else
(auto-suggest system) wasn't ready when the user wants to use a custom
expression.

With this change, we immediately prime the parser whenever a query
builder is started, under the assumption that the app isn't being
embedded (hence, past the concern address in PR #14133) and from this
point onward, the custom expression parser etc will be likely needed.