Upgrade React to v16 (#14391)
* Upgrade React to v16

* Upgrade react-hot-loader

* Use portal in BodyComponent (#14779)

* Use portal in Popover (#14799)

* Use portal in Popover

* Use SandboxedPortal to stop portal child/parent event propagation

React portals bubble events from component childrent to component
parents _through_ the portal as though they are attached. This breaks
our Triggerable HOC that wraps everything in an anchor tag -- click
events inside of an open Popover trigger the "toggle" function in
Triggerable, closing the Popover component.

Adding an additional div to the code potentially breaks styling, but
this seemed better than the alternative of changing/limiting
Triggerable.

* Fix adding of open class

* don't trigger some logic when popover is not open

* change wrapper span to a div

* rmv explicit bind from handleDismissal fn

* add comment explaining reasoning for SandboxedPortal

* add comment to explain need for 'maxHeight' state value

* Use ReactDOMServer to render html in a contenteditable div (#14869)

As part of the react v16 we can no longer rely on a synchronous
ReactDOM.render call. Using the third arg callback of ReactDOM.render
does not work for reasons unknown to me, but it would likely be
race-condition-y anyways due to this being an input field. I'm using
ReactDOMServer in order to avoid rewriting the react TokenizedExpression
component as html and to avoid associated xss vulnerabilities, etc.

* Rename unsafe lifecycles (#14887)

* Rename unsafe lifecycles

* Fix flow error: Remove unused suppressions

* use SandboxedPortal in Modal (#14905)

* use SandboxedPortal in Modal

* Move modal append container to constructor

This breaks the entry animation but it fixes problems with having a
popover _inside_ of the modal. Previously, popover was appending its
container el before the modal, causing a stacking issue.

* remove unstable_renderSubtreeIntoContainer from Tooltip (#14976)

Co-authored-by: Dalton <daltojohnso@users.noreply.github.com>