drop columns that don't have view_as
so we don't overwrite the top level stuff if there is already a
click_behavior. That signifies that either this migration has run, or
perhaps already has had manual intervention. An open question is if we
extend this to the columns: if there's manual intervention do we
really want to be mucking about in here?

Also, go back to the dropping columns style.

```clojure
(if (and ...)
    (assoc m col (merge field-settings click-behavior))
    m) ;; here we omit the column again in our reduce-kv
```