Refactor ParameterValueWidget
Primary goal is to be more precise with what props we pass to the
"Widget" component as we are willy-nilly passing props when we don't
need to, and I'd prefer to be explicit (and correct) with which props we
pass to Widget.

Also pulled a few pure functions off the component and replaced the
deprecated lifecycle functions we were using.