Don't merge existing fix from card. can't happen
this change though did help out. we want to resolve when there are old
style and new style on the same dashcard. In this case, the new style
should persist:

```clojure
(testing "If there is migration eligible on dash but also new style on dash, new style wins"
  (let [dash {"column_settings"
              {"[\"ref\",[\"field-id\",4]]"
               {"view_as"       "link"
                "link_template" "http://old" ;; this stuff could be migrated
                "link_text"     "old"
                "column_title"  "column title"
                "click_behavior"
                {"type" "link",
                 "linkType" "url", ;; but there is already a new style and it wins
                 "linkTemplate" "http://new",
                 "linkTextTemplate" "new"}}}}]
    (is (= {"column_settings"
            {"[\"ref\",[\"field-id\",4]]"
             {"click_behavior"
              {"type" "link",
               "linkType" "url",
               "linkTemplate" "http://new",
               "linkTextTemplate" "new"},
              "column_title" "column title"}}}
           (migrate nil dash)))))
```

note the column settings for field-id 4 have the old style and the new
style but the result is that the new style survives. This happens from
the `(m/deep-merge fixed-card fixed-dashcard (existing-fixed
dashcard))` as the existing is put on top